extracted type definitions
- deleted file to be reworked (deletingM2AItems.ts)
